Update 06_FlowControl.Rmd
Some grammar and spelling.

Edited `if else` for consistency throughout the chapter, so that it is always in monospaced font.

In the Exercises, 2(b)(iii) does not have any explicit instructions. The instructions in 2(b) refer to the next two parts, (I) and (ii), but not (iii), so the reader might be confused and think they need to replace these ??? with `&` or `|`, instead of with one of the output values. I edited line 479 to reflect `<=` instead of `<` in the appropriate places. It might also make sense in Exercise 2 in general to have the value set for `b` to be greater than 10, so that all conditions in the `if else` statements are obtainable. But, the enterprising reader may make this change on their own regardless.

In line 219, the `while` loop condition is evaluating based on the random normal sample `x` from line 162, which is a vector of length 20. If the desired result is a terminating loop without error, a new value of `x` should be set just prior to the `while` loop